+++ mpich2/test/mpi/rma/putfence1.c
@@ -49,7 +49,7 @@ int main( int argc, char *argv[] )
 
 		MPI_Type_extent( recvtype.datatype, &extent );
 		MPI_Win_create( recvtype.buf, recvtype.count * extent, 
-				(int)extent, MPI_INFO_NULL, comm, &win );
+				extent, MPI_INFO_NULL, comm, &win );
 		MPI_Win_fence( 0, win );
 		if (rank == source) {
 		    /* To improve reporting of problems about operations, we
